[GEOT-6878] Avoid unnecessary object wrapper creation, add a PMD rule to check it

+++ modules/library/main/src/main/java/org/geotools/feature/visitor/QuantileListVisitor.java
@@ -69,7 +69,7 @@ public class QuantileListVisitor implements FeatureCalc {
         }
 
         // calculate number of items to put into each of the larger bins
-        int binPop = Double.valueOf(Math.ceil((double) count / bins)).intValue();
+        int binPop = (int) (Math.ceil((double) count / bins));
         // determine index of bin where the next bin has one less item
         int lastBigBin = count % bins;
         if (lastBigBin == 0) lastBigBin = bins;
